+++ checkstyle/import-control-core.xml
@@ -111,5 +111,13 @@
       <allow pkg="org.apache.kafka.clients"/>
       <allow pkg="org.apache.kafka.metadata" />
     </subpackage>
+    <subpackage name="zk">
+      <allow pkg="io.netty"/>
+      <allow pkg="kafka.cluster"/>
+      <allow pkg="kafka.network"/>
+      <allow pkg="kafka.zookeeper"/>
+      <allow pkg="org.apache.zookeeper"/>
+    </subpackage>
   </subpackage>
+
 </import-control>
